Atirando e destruindo inimigos

+++ src/com/trigger/entity/BulletShoot.java
@@ -8,12 +8,12 @@ import java.awt.image.BufferedImage;
 
 public class BulletShoot extends Entity{
 
-    private int dx, dy;
+    private double dx, dy;
     private double speed = 2;
 
     private int life = 40, curLife = 0;
 
-    public BulletShoot(int x, int y, int width, int height, BufferedImage sprite, int dx, int dy) {
+    public BulletShoot(int x, int y, int width, int height, BufferedImage sprite, double dx, double dy) {
         super(x, y, width, height, sprite);
         this.dx = dx;
         this.dy = dy;
